move display devices to folder
