+++ src/intro.py
@@ -1,6 +1,6 @@
 from PIL import Image, ImageDraw
 import time
-from .image_utils import draw_centered_text
+from .image_utils import centered_text
 from .network_utils import wait_for_internet_connection
 page1 = '''Hi, I'm your new Bitbot.
 
@@ -56,7 +56,7 @@ class Intro:
         background = self.get_background()
         img = Image.new("RGBA", self.display_size, transparent)
         draw = ImageDraw.Draw(img)
-        draw_centered_text(draw, page1, self.font, img.size, 'topleft')
+        centered_text(draw, page1, self.font, img.size, 'topleft')
         background.paste(img, (0, 0), img)
         yield background
 
@@ -65,7 +65,7 @@ class Intro:
         img = Image.new("RGBA", self.display_size, transparent)
         img.paste(background)
         draw = ImageDraw.Draw(img)
-        draw_centered_text(draw, page2, self.font, img.size, 'topleft')
+        centered_text(draw, page2, self.font, img.size, 'topleft')
         background.paste(img, (0, 0), img)
         yield background
 
@@ -74,7 +74,7 @@ class Intro:
         img = Image.new("RGBA", self.display_size, transparent)
         img.paste(background)
         draw = ImageDraw.Draw(img)
-        draw_centered_text(draw, page3, self.font, img.size, 'topleft')
+        centered_text(draw, page3, self.font, img.size, 'topleft')
         background.paste(img, (0, 0), img)
         yield background
