+++ tests/test_device_picker.py
@@ -1,5 +1,5 @@
 import unittest
-import src.display.kinky as kinky
+from src.display.picker import picker, disk, inky, waveshare
 from os.path import join as pjoin
 import pathlib
 from PIL import Image
@@ -13,23 +13,23 @@ config = load_config_ini(files)
 
 def get_device(display_output_name):
     config.set('display', 'output', display_output_name)
-    return kinky.picker(config)
+    return picker(config)
 
 
 class TestDevicePicker(unittest.TestCase):
     def test_can_create_disker(self):
         device = get_device('disk')
-        assert type(device) is kinky.Disker, "should pick disk output"
+        assert type(device) is disk.Disker, "should pick disk output"
 
     @unittest.skip("inky device must be connected")
     def test_can_create_inky(self):
         device = get_device('inky')
-        assert type(device) is kinky.Inker, "should pick inky output"
+        assert type(device) is inky.Inker, "should pick inky output"
 
     @unittest.skip("waveshare device must be connected")
     def test_can_create_waveshare(self):
         device = get_device('waveshare.epd2in7b_V2')
-        assert type(device) is kinky.Waver, "should pick waveshare output"
+        assert type(device) is waveshare.Waver, "should pick waveshare output"
 
     @unittest.skip("waveshare device must be connected")
     def test_draw_image_to_waveshare(self):
